@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 30
 
     defaultConfig {
         applicationId "com.mxn.soul.flowingdrawer"
-        minSdkVersion 11
-        targetSdkVersion 26
+        minSdkVersion 14
+        targetSdkVersion 30
         versionCode 120
         versionName "1.2.0"
     }
@@ -20,12 +20,12 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:22.2.0'
-    implementation 'com.android.support:recyclerview-v7:22.+'
-    implementation 'com.android.support:design:22.2.1'
     implementation 'com.squareup.picasso:picasso:2.5.0'
-    implementation 'com.android.support:cardview-v7:22.+'
     implementation 'com.nineoldandroids:library:2.4.0'
     implementation 'com.mxn.soul:flowingdrawer-core:2.1.0'
+    implementation 'androidx.coordinatorlayout:coordinatorlayout:1.1.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'com.google.android.material:material:1.2.1'
+    implementation 'androidx.cardview:cardview:1.0.0'
 //    implementation project(':flowingdrawer-core')
 }
